tpm-tools: convert all cpp to c files
Drop the dependency on C++ as we were not really using
any compelling language features.
