+++ src/tpm2_create.c
@@ -37,17 +37,17 @@
 #include <limits.h>
 #include <ctype.h>
 #include <getopt.h>
+#include <stdbool.h>
 
 #include <sapi/tpm20.h>
 #include <tcti/tcti_socket.h>
 #include "common.h"
 
-int debugLevel = 0;
 TPMS_AUTH_COMMAND sessionData;
 bool hexPasswd = false;
-TPM_HANDLE handle2048rsa;
+int debugLevel = 0;
 
-int setAlg(TPMI_ALG_PUBLIC type,TPMI_ALG_HASH nameAlg,TPM2B_PUBLIC *inPublic)
+int setAlg(TPMI_ALG_PUBLIC type,TPMI_ALG_HASH nameAlg,TPM2B_PUBLIC *inPublic, int I_flag)
 {
     switch(nameAlg)
     {
@@ -66,21 +66,19 @@ int setAlg(TPMI_ALG_PUBLIC type,TPMI_ALG_HASH nameAlg,TPM2B_PUBLIC *inPublic)
 
     // First clear attributes bit field.
     *(UINT32 *)&(inPublic->t.publicArea.objectAttributes) = 0;
-    inPublic->t.publicArea.objectAttributes.restricted = 1;
+    inPublic->t.publicArea.objectAttributes.restricted = 0;
     inPublic->t.publicArea.objectAttributes.userWithAuth = 1;
     inPublic->t.publicArea.objectAttributes.decrypt = 1;
+    inPublic->t.publicArea.objectAttributes.sign = 1;
     inPublic->t.publicArea.objectAttributes.fixedTPM = 1;
     inPublic->t.publicArea.objectAttributes.fixedParent = 1;
     inPublic->t.publicArea.objectAttributes.sensitiveDataOrigin = 1;
-    inPublic->t.publicArea.authPolicy.t.size = 0;
 
     inPublic->t.publicArea.type = type;
     switch(type)
     {
     case TPM_ALG_RSA:
-        inPublic->t.publicArea.parameters.rsaDetail.symmetric.algorithm = TPM_ALG_AES;
-        inPublic->t.publicArea.parameters.rsaDetail.symmetric.keyBits.aes = 128;
-        inPublic->t.publicArea.parameters.rsaDetail.symmetric.mode.aes = TPM_ALG_CFB;
+        inPublic->t.publicArea.parameters.rsaDetail.symmetric.algorithm = TPM_ALG_NULL;
         inPublic->t.publicArea.parameters.rsaDetail.scheme.scheme = TPM_ALG_NULL;
         inPublic->t.publicArea.parameters.rsaDetail.keyBits = 2048;
         inPublic->t.publicArea.parameters.rsaDetail.exponent = 0;
@@ -88,16 +86,26 @@ int setAlg(TPMI_ALG_PUBLIC type,TPMI_ALG_HASH nameAlg,TPM2B_PUBLIC *inPublic)
         break;
 
     case TPM_ALG_KEYEDHASH:
-        inPublic->t.publicArea.parameters.keyedHashDetail.scheme.scheme = TPM_ALG_XOR;
-        inPublic->t.publicArea.parameters.keyedHashDetail.scheme.details.exclusiveOr.hashAlg = TPM_ALG_SHA256;
-        inPublic->t.publicArea.parameters.keyedHashDetail.scheme.details.exclusiveOr.kdf = TPM_ALG_KDF1_SP800_108;
         inPublic->t.publicArea.unique.keyedHash.t.size = 0;
+        inPublic->t.publicArea.objectAttributes.decrypt = 0;
+        if (I_flag)
+        {
+            // sealing
+            inPublic->t.publicArea.objectAttributes.sign = 0;
+            inPublic->t.publicArea.objectAttributes.sensitiveDataOrigin = 0;
+            inPublic->t.publicArea.parameters.keyedHashDetail.scheme.scheme = TPM_ALG_NULL;
+        }
+        else
+        {
+            // hmac
+            inPublic->t.publicArea.objectAttributes.sign = 1;
+            inPublic->t.publicArea.parameters.keyedHashDetail.scheme.scheme = TPM_ALG_HMAC;
+            inPublic->t.publicArea.parameters.keyedHashDetail.scheme.details.hmac.hashAlg = nameAlg;  //for tpm2_hmac multi alg
+        }
         break;
 
     case TPM_ALG_ECC:
-        inPublic->t.publicArea.parameters.eccDetail.symmetric.algorithm = TPM_ALG_AES;
-        inPublic->t.publicArea.parameters.eccDetail.symmetric.keyBits.aes = 128;
-        inPublic->t.publicArea.parameters.eccDetail.symmetric.mode.sym = TPM_ALG_CFB;
+        inPublic->t.publicArea.parameters.eccDetail.symmetric.algorithm = TPM_ALG_NULL;
         inPublic->t.publicArea.parameters.eccDetail.scheme.scheme = TPM_ALG_NULL;
         inPublic->t.publicArea.parameters.eccDetail.curveID = TPM_ECC_NIST_P256;
         inPublic->t.publicArea.parameters.eccDetail.kdf.scheme = TPM_ALG_NULL;
@@ -119,9 +127,9 @@ int setAlg(TPMI_ALG_PUBLIC type,TPMI_ALG_HASH nameAlg,TPM2B_PUBLIC *inPublic)
     return 0;
 }
 
-int createPrimary(TPMI_RH_HIERARCHY hierarchy, TPM2B_PUBLIC *inPublic, TPM2B_SENSITIVE_CREATE *inSensitive, TPMI_ALG_PUBLIC type, TPMI_ALG_HASH nameAlg, int P_flag, int K_flag)
+int create(TPMI_DH_OBJECT parentHandle, TPM2B_PUBLIC *inPublic, TPM2B_SENSITIVE_CREATE *inSensitive, TPMI_ALG_PUBLIC type, TPMI_ALG_HASH nameAlg, const char *opuFilePath, const char *oprFilePath, int o_flag, int O_flag, int I_flag, int A_flag, UINT32 objectAttributes)
 {
-    UINT32 rval;
+    TPM_RC rval;
     TPMS_AUTH_RESPONSE sessionDataOut;
     TSS2_SYS_CMD_AUTHS sessionsData;
     TSS2_SYS_RSP_AUTHS sessionsDataOut;
@@ -130,8 +138,9 @@ int createPrimary(TPMI_RH_HIERARCHY hierarchy, TPM2B_PUBLIC *inPublic, TPM2B_SEN
 
     TPM2B_DATA              outsideInfo = { { 0, } };
     TPML_PCR_SELECTION      creationPCR;
-    TPM2B_NAME              name = { { sizeof(TPM2B_NAME)-2, } };
     TPM2B_PUBLIC            outPublic = { { 0, } };
+    TPM2B_PRIVATE           outPrivate = { { sizeof(TPM2B_PRIVATE)-2, } };
+
     TPM2B_CREATION_DATA     creationData = { { 0, } };
     TPM2B_DIGEST            creationHash = { { sizeof(TPM2B_DIGEST)-2, } };
     TPMT_TK_CREATION        creationTicket = { 0, };
@@ -142,16 +151,13 @@ int createPrimary(TPMI_RH_HIERARCHY hierarchy, TPM2B_PUBLIC *inPublic, TPM2B_SEN
     sessionsDataOut.rspAuths = &sessionDataOutArray[0];
     sessionsData.cmdAuths = &sessionDataArray[0];
 
-    sessionsData.cmdAuthsCount = 1;
     sessionsDataOut.rspAuthsCount = 1;
 
     sessionData.sessionHandle = TPM_RS_PW;
     sessionData.nonce.t.size = 0;
 
-    if(P_flag == 0)
-        sessionData.hmac.t.size = 0;
-
-    *((UINT8 *)((void *)&sessionData.sessionAttributes)) = 0;
+    sessionsData.cmdAuthsCount = 1;
+    sessionsData.cmdAuths[0] = &sessionData;
     if (sessionData.hmac.t.size > 0 && hexPasswd)
     {
         sessionData.hmac.t.size = sizeof(sessionData.hmac) - 2;
@@ -159,13 +165,11 @@ int createPrimary(TPMI_RH_HIERARCHY hierarchy, TPM2B_PUBLIC *inPublic, TPM2B_SEN
                               &sessionData.hmac.t.size,
                               sessionData.hmac.t.buffer) != 0)
         {
-            printf( "Failed to convert Hex format password for hierarchy Passwd.\n");
+            printf( "Failed to convert Hex format password for parent Passwd.\n");
             return -1;
         }
     }
 
-    if(K_flag == 0)
-        inSensitive->t.sensitive.userAuth.t.size = 0;
     if (inSensitive->t.sensitive.userAuth.t.size > 0 && hexPasswd)
     {
         inSensitive->t.sensitive.userAuth.t.size = sizeof(inSensitive->t.sensitive.userAuth) - 2;
@@ -173,55 +177,73 @@ int createPrimary(TPMI_RH_HIERARCHY hierarchy, TPM2B_PUBLIC *inPublic, TPM2B_SEN
                               &inSensitive->t.sensitive.userAuth.t.size,
                               inSensitive->t.sensitive.userAuth.t.buffer) != 0)
         {
-            printf( "Failed to convert Hex format password for primary Passwd.\n");
+            printf( "Failed to convert Hex format password for object Passwd.\n");
             return -1;
         }
     }
-
-    inSensitive->t.sensitive.data.t.size = 0;
     inSensitive->t.size = inSensitive->t.sensitive.userAuth.b.size + 2;
 
-    if(setAlg(type, nameAlg, inPublic))
+    if(setAlg(type, nameAlg, inPublic, I_flag))
         return -1;
+
+    if(A_flag == 1)
+        inPublic->t.publicArea.objectAttributes.val = objectAttributes;
+    printf("ObjectAttribute: 0x%08X\n",inPublic->t.publicArea.objectAttributes.val);
 
     creationPCR.count = 0;
 
-    rval = Tss2_Sys_CreatePrimary(sysContext, hierarchy, &sessionsData, inSensitive, inPublic,  &outsideInfo, &creationPCR, &handle2048rsa, &outPublic, &creationData, &creationHash, &creationTicket, &name, &sessionsDataOut);
+    rval = Tss2_Sys_Create(sysContext, parentHandle, &sessionsData, inSensitive, inPublic,
+            &outsideInfo, &creationPCR, &outPrivate,&outPublic,&creationData, &creationHash,
+            &creationTicket, &sessionsDataOut);
     if(rval != TPM_RC_SUCCESS)
     {
-        printf("\nCreatePrimary Failed ! ErrorCode: 0x%0x\n\n",rval);
+        printf("\nCreate Object Failed ! ErrorCode: 0x%0x\n\n",rval);
         return -2;
     }
-    printf("\nCreatePrimary Succeed ! Handle: 0x%8.8x\n\n",handle2048rsa);
+    printf("\nCreate Object Succeed !\n");
+
+    if(o_flag == 1)
+    {
+        if(saveDataToFile(opuFilePath, (UINT8 *)&outPublic, sizeof(outPublic)))
+            return -3;
+    }
+    if(O_flag == 1)
+    {
+        if(saveDataToFile(oprFilePath, (UINT8 *)&outPrivate, sizeof(outPrivate)))
+            return -4;
+    }
+
     return 0;
 }
 
 void showHelp(const char *name)
 {
-    printf("\n%s [options]\n"
+    printf("\n%s  [options]\n"
         "\n"
-        "-h, --help               Display command tool usage info;\n"
-        "-v, --version            Display command tool version info\n"
-        "-A, --auth <o | p |...>  the authorization used to authorize thecommands\n"
-            "\to  TPM_RH_OWNER\n"
-            "\tp  TPM_RH_PLATFORM\n"
-            "\te  TPM_RH_ENDORSEMENT\n"
-            "\tn  TPM_RH_NULL\n"
-        "-P, --pwdp <string>      password for hierarchy, optional\n"
-        "-K, --pwdk <string>      password for key, optional\n"
-        "-g, --halg <sha1,...>    algorithm used for computing the Name of the object\n"
+        "-h, --help             Display command tool usage info;\n"
+        "-v, --version          Display command tool version info\n"
+        "-H, --parent <handle>  parent handle\n"
+        "-c, --contextParent <filename> filename for parent context\n"
+        "-P, --pwdp   <string>  password for parent key, optional\n"
+        "-K, --pwdk   <string>  password for key, optional\n"
+        "-g, --halg   <hexAlg>  algorithm used for computing the Name of the object\n"
             "\t0x0004  TPM_ALG_SHA1\n"
             "\t0x000B  TPM_ALG_SHA256\n"
             "\t0x000C  TPM_ALG_SHA384\n"
             "\t0x000D  TPM_ALG_SHA512\n"
             "\t0x0012  TPM_ALG_SM3_256\n"
-        "-G, --kalg <rsa,...>     algorithm associated with this object\n"
+        "-G, --kalg   <hexAlg>  algorithm associated with this object\n"
             "\t0x0001  TPM_ALG_RSA\n"
             "\t0x0008  TPM_ALG_KEYEDHASH\n"
             "\t0x0023  TPM_ALG_ECC\n"
             "\t0x0025  TPM_ALG_SYMCIPHER\n"
-        "-C, --context <filename> The file to save the object context, optional\n"
-        "-X, --passwdInHex        passwords given by any options are hex format.\n"
+        "-A, --objectAttribute <hexAttributeDWord>  object attributes, optional\n"
+        "-I, --inFile          <dataFileToBeSealed>  data file to be sealed, optional\n"
+        "-L, --pol <policyFile>         the input policy file, optional\n"
+        "-o, --opu <publicKeyFileName>  the output file which contains the public key, optional\n"
+        "-O, --opr <privateKeyFileName> the output file which contains the private key, optional\n"
+        "-X, --passwdInHex              passwords given by any options are hex format.\n"
+
         "-p, --port  <port number>  The Port number, default is %d, optional\n"
         "-d, --debugLevel <0|1|2|3> The level of debug message, default is 0, optional\n"
             "\t0 (high level test results)\n"
@@ -230,10 +252,10 @@ void showHelp(const char *name)
             "\t3 (resource manager tables)\n"
         "\n"
         "Example:\n"
-        "%s -A e -P abc123 -K def456 -g 0x000B -G 0x0008 \n"
-        "%s -A p -g 0x000B -G 0x0008\n\n"//-i <simulator IP>\n\n"
-        "%s -A e -P 1a1b1c -K 456def -X -g 0x000B -G 0x0008 \n"
-        , name, DEFAULT_RESMGR_TPM_PORT, name, name, name);
+        "%s -H 0x80000000 -P abc123 -K def456 -g 0x000B -G 0x0008 -I data.File -o opu.File\n"
+        "%s -H 0x80000000 -g 0x000B -G 0x0008 -I data.File -o opu.File -O opr.File\n\n"// -i <simulator IP>\n\n",DEFAULT_TPM_PORT);
+        "%s -H 0x80000000 -P 123abc -K 456def -X -g 0x000B -G 0x0008 -I data.File -o opu.File\n"
+        ,name, DEFAULT_RESMGR_TPM_PORT, name, name, name);
 }
 
 int main(int argc, char* argv[])
@@ -243,28 +265,38 @@ int main(int argc, char* argv[])
     int port = DEFAULT_RESMGR_TPM_PORT;
 
     TPM2B_SENSITIVE_CREATE  inSensitive;
+    inSensitive.t.sensitive.data.t.size = 0;
     TPM2B_PUBLIC            inPublic;
     TPMI_ALG_PUBLIC type;
     TPMI_ALG_HASH nameAlg;
-    TPMI_RH_HIERARCHY hierarchy = TPM_RH_NULL;
+    TPMI_DH_OBJECT parentHandle;
+    UINT32 objectAttributes = 0;
+    char opuFilePath[PATH_MAX] = {0};
+    char oprFilePath[PATH_MAX] = {0};
+    char *contextParentFilePath = NULL;
 
     setbuf(stdout, NULL);
     setvbuf (stdout, NULL, _IONBF, BUFSIZ);
 
     int opt = -1;
-    const char *optstring = "hvA:P:K:g:G:p:d:C:X";
+    const char *optstring = "hvH:P:K:g:G:A:I:L:o:O:p:d:c:X";
     static struct option long_options[] = {
       {"help",0,NULL,'h'},
       {"version",0,NULL,'v'},
-      {"auth",1,NULL,'A'},
+      {"parent",1,NULL,'H'},
       {"pwdp",1,NULL,'P'},
       {"pwdk",1,NULL,'K'},
       {"halg",1,NULL,'g'},
       {"kalg",1,NULL,'G'},
+      {"objectAttributes",1,NULL,'A'},
+      {"inFile",1,NULL,'I'},
+      {"policyFile",1,NULL,'L'},
+      {"opu",1,NULL,'o'},
+      {"opr",1,NULL,'O'},
+      {"contextParent",1,NULL,'c'},
+      {"passwdInHex",0,NULL,'X'},
       {"port",1,NULL,'p'},
       {"debugLevel",1,NULL,'d'},
-      {"context",1,NULL,'C'},
-      {"passwdInHex",0,NULL,'X'},
       {0,0,0,0}
     };
 
@@ -273,14 +305,18 @@ int main(int argc, char* argv[])
     int flagCnt = 0;
     int h_flag = 0,
         v_flag = 0,
-        A_flag = 0,
+        H_flag = 0,
         P_flag = 0,
         K_flag = 0,
         g_flag = 0,
         G_flag = 0,
-        C_flag = 0;
-    char *contextFile = NULL;
-
+        A_flag = 0,
+        I_flag = 0,
+        L_flag = 0,
+        o_flag = 0,
+        c_flag = 0,
+        O_flag = 0/*,
+        f_flag = 0*/;
     if(argc == 1)
     {
         showHelp(argv[0]);
@@ -297,21 +333,14 @@ int main(int argc, char* argv[])
         case 'v':
             v_flag = 1;
             break;
-        case 'A':
-            if(strcmp(optarg,"o") == 0 || strcmp(optarg,"O") == 0)
-                hierarchy = TPM_RH_OWNER;
-            else if(strcmp(optarg,"p") == 0 || strcmp(optarg,"P") == 0)
-                hierarchy = TPM_RH_PLATFORM;
-            else if(strcmp(optarg,"e") == 0 || strcmp(optarg,"E") == 0)
-                hierarchy = TPM_RH_ENDORSEMENT;
-            else if(strcmp(optarg,"n") == 0 || strcmp(optarg,"N") == 0)
-                hierarchy = TPM_RH_NULL;
-            else
+        case 'H':
+            if(getSizeUint32Hex(optarg,&parentHandle) != 0)
             {
+                showArgError(optarg, argv[0]);
                 returnVal = -1;
                 break;
             }
-            A_flag = 1;
+            H_flag = 1;
             break;
 
         case 'P':
@@ -321,7 +350,6 @@ int main(int argc, char* argv[])
                 returnVal = -2;
                 break;
             }
-
             P_flag = 1;
             break;
         case 'K':
@@ -353,40 +381,86 @@ int main(int argc, char* argv[])
             printf("type = 0x%4.4x\n", type);
             G_flag = 1;
             break;
+        case 'A':
+            if(getSizeUint32Hex(optarg,&objectAttributes) != 0)
+            {
+                showArgError(optarg, argv[0]);
+                returnVal = -6;
+                break;
+            }
+            A_flag = 1;//H_flag = 1;
+            break;
+        case 'I':
+            inSensitive.t.sensitive.data.t.size = sizeof(inSensitive.t.sensitive.data) - 2;
+            if(loadDataFromFile(optarg, inSensitive.t.sensitive.data.t.buffer, &inSensitive.t.sensitive.data.t.size) != 0)
+            {
+                returnVal = -7;
+                break;
+            }
+            I_flag = 1;
+            printf("inSensitive.t.sensitive.data.t.size = %d\n",inSensitive.t.sensitive.data.t.size);
+            break;
+        case 'L':
+            inPublic.t.publicArea.authPolicy.t.size = sizeof(inPublic.t.publicArea.authPolicy) - 2;
+            if(loadDataFromFile(optarg, inPublic.t.publicArea.authPolicy.t.buffer, &inPublic.t.publicArea.authPolicy.t.size) != 0)
+            {
+                returnVal = -8;
+                break;
+            }
+            L_flag = 1;
+            break;
+        case 'o':
+            safeStrNCpy(opuFilePath, optarg, sizeof(opuFilePath));
+            if(checkOutFile(opuFilePath) != 0)
+            {
+                returnVal = -9;
+                break;
+            }
+            o_flag = 1;
+            break;
+        case 'O':
+            safeStrNCpy(oprFilePath, optarg, sizeof(oprFilePath));
+            if(checkOutFile(oprFilePath) != 0)
+            {
+                returnVal = -10;
+                break;
+            }
+            O_flag = 1;
+            break;
+        case 'c':
+            contextParentFilePath = optarg;
+            if(contextParentFilePath == NULL || contextParentFilePath[0] == '\0')
+            {
+                returnVal = -11;
+                break;
+            }
+            printf("contextParentFile = %s\n", contextParentFilePath);
+            c_flag = 1;
+            break;
+        case 'X':
+            hexPasswd = true;
+            break;
         case 'p':
             if( getPort(optarg, &port) )
             {
                 printf("Incorrect port number.\n");
-                returnVal = -6;
+                returnVal = -12;
             }
             break;
         case 'd':
             if( getDebugLevel(optarg, &debugLevel) )
             {
                 printf("Incorrect debug level.\n");
-                returnVal = -7;
+                returnVal = -13;
             }
-            break;
-        case 'C':
-            contextFile = optarg;
-            if(contextFile == NULL || contextFile[0] == '\0')
-            {
-                returnVal = -8;
-                break;
-            }
-            printf("contextFile = %s\n", contextFile);
-            C_flag = 1;
-            break;
-        case 'X':
-            hexPasswd = true;
             break;
         case ':':
 //              printf("Argument %c needs a value!\n",optopt);
-            returnVal = -9;
+            returnVal = -14;
             break;
         case '?':
 //              printf("Unknown Argument: %c\n",optopt);
-            returnVal = -10;
+            returnVal = -15;
             break;
         //default:
         //  break;
@@ -397,7 +471,19 @@ int main(int argc, char* argv[])
 
     if(returnVal != 0)
         return returnVal;
-    flagCnt = h_flag + v_flag + A_flag + g_flag + G_flag;
+
+    if(P_flag == 0)
+        sessionData.hmac.t.size = 0;
+    if(I_flag == 0)
+        inSensitive.t.sensitive.data.t.size = 0;
+    if(K_flag == 0)
+        inSensitive.t.sensitive.userAuth.t.size = 0;
+    if(L_flag == 0)
+        inPublic.t.publicArea.authPolicy.t.size = 0;
+
+    *((UINT8 *)((void *)&sessionData.sessionAttributes)) = 0;
+
+    flagCnt = h_flag + v_flag + H_flag + g_flag + G_flag + c_flag ;
     if(flagCnt == 1)
     {
         if(h_flag == 1)
@@ -407,29 +493,27 @@ int main(int argc, char* argv[])
         else
         {
             showArgMismatch(argv[0]);
-            return -11;
+            return -16;
         }
     }
-    else if(flagCnt == 3 && A_flag == 1 && g_flag == 1 && G_flag == 1)
+    else if(flagCnt == 3 && (H_flag == 1 || c_flag == 1) && g_flag == 1 && G_flag == 1)
     {
-
         prepareTest(hostName, port, debugLevel);
 
-        returnVal = createPrimary(hierarchy, &inPublic, &inSensitive, type, nameAlg, P_flag, K_flag);
-
-        if (returnVal == 0 && C_flag)
-            returnVal = saveTpmContextToFile(sysContext, handle2048rsa, contextFile);
+        if(c_flag)
+            returnVal = loadTpmContextFromFile(sysContext, &parentHandle, contextParentFilePath);
+        if(returnVal == 0)
+            returnVal = create(parentHandle, &inPublic, &inSensitive, type, nameAlg, opuFilePath, oprFilePath, o_flag, O_flag, I_flag, A_flag, objectAttributes);
 
         finishTest();
 
         if(returnVal)
-            return -12;
+            return -17;
     }
     else
     {
         showArgMismatch(argv[0]);
-        return -13;
+        return -18;
     }
-
     return 0;
 }
