+++ src/tpm2_loadexternal.c
