+++ src/tpm2_makecredential.c
