+++ src/tpm2_nvdefine.c
@@ -30,6 +30,7 @@
 //**********************************************************************;
 
 #include <stdarg.h>
+#include <stdbool.h>
 
 #include <stdlib.h>
 #include <stdio.h>
@@ -45,21 +46,21 @@
 int debugLevel = 0;
 UINT32 nvIndex = 0;
 UINT32 authHandle = TPM_RH_PLATFORM;
-UINT16 dataSize = 0;
-unsigned char nvBuffer[MAX_NV_INDEX_SIZE];
-char fileName[PATH_MAX];
+UINT32 size = 0;
+UINT32 nvAttribute = 0;
 char handlePasswd[sizeof(TPMU_HA)];
+char indexPasswd[sizeof(TPMU_HA)];
 bool hexPasswd = false;
 
-int nvWrite()
+int nvSpaceDefine()
 {
     UINT32 rval;
+    TPM2B_AUTH nvAuth;
+    TPM2B_NV_PUBLIC publicInfo;
     TPMS_AUTH_COMMAND sessionData;
     TPMS_AUTH_RESPONSE sessionDataOut;
     TSS2_SYS_CMD_AUTHS sessionsData;
     TSS2_SYS_RSP_AUTHS sessionsDataOut;
-    UINT16 i, offset = 0;
-    TPM2B_MAX_NV_BUFFER nvWriteData;
 
     TPMS_AUTH_COMMAND *sessionDataArray[1] = { &sessionData };
     TPMS_AUTH_RESPONSE *sessionDataOutArray[1] = { &sessionDataOut };
@@ -91,28 +92,40 @@ int nvWrite()
         }
     }
 
-    while (dataSize > 0)
-    {
-        nvWriteData.t.size = dataSize > MAX_NV_BUFFER_SIZE ? MAX_NV_BUFFER_SIZE : dataSize;
-        printf("\nThe data(size=%d) to be written:\n", nvWriteData.t.size);
-        for( i = 0; i<(int)nvWriteData.t.size; i++ )
-        {
-            nvWriteData.t.buffer[i] = nvBuffer[offset+i];
-            printf("%02x ", nvBuffer[offset+i]);
-        }
-        printf("\n\n");
+    publicInfo.t.size = sizeof( TPMI_RH_NV_INDEX ) +
+            sizeof( TPMI_ALG_HASH ) + sizeof( TPMA_NV ) + sizeof( UINT16) +
+            sizeof( UINT16 );
+    publicInfo.t.nvPublic.nvIndex = nvIndex;
+    publicInfo.t.nvPublic.nameAlg = TPM_ALG_SHA256;
 
-        rval = Tss2_Sys_NV_Write( sysContext, authHandle, nvIndex, &sessionsData, &nvWriteData, offset, &sessionsDataOut );
-        if(rval != TSS2_RC_SUCCESS)
+    // Now set the attributes.
+    publicInfo.t.nvPublic.attributes.val = nvAttribute;
+    publicInfo.t.nvPublic.authPolicy.t.size = 0;
+    publicInfo.t.nvPublic.dataSize = size;
+
+    nvAuth.t.size = 0;
+    if (strlen(indexPasswd) > 0 && !hexPasswd)
+    {
+        nvAuth.t.size = strlen(indexPasswd);
+        memcpy( &nvAuth.t.buffer[0], indexPasswd, nvAuth.t.size );
+    }
+    else if (strlen(indexPasswd) > 0 && hexPasswd)
+    {
+        nvAuth.t.size = sizeof(nvAuth) - 2;
+        if (hex2ByteStructure(indexPasswd,&nvAuth.t.size,nvAuth.t.buffer) != 0)
         {
-            printf( "Failed to write NV area at index 0x%x (%d) offset 0x%x. Error:0x%x\n", nvIndex, nvIndex, offset, rval );
+            printf( "Failed to convert Hex format password for indexPasswd.\n");
             return -1;
         }
-        printf( "Success to write NV area at index 0x%x (%d) offset 0x%x.\n", nvIndex, nvIndex, offset );
-
-        dataSize -= nvWriteData.t.size;
-        offset += nvWriteData.t.size;        
     }
+
+    rval = Tss2_Sys_NV_DefineSpace( sysContext, authHandle, &sessionsData, &nvAuth, &publicInfo, &sessionsDataOut );
+    if(rval != TPM_RC_SUCCESS)
+    {
+        printf( "\nFailed to define NV area at index 0x%x (%d).Error:0x%x\n", nvIndex, nvIndex, rval );
+        return -1;
+    }
+    printf( "Success to define NV area at index 0x%x (%d).\n", nvIndex, nvIndex);
 
     return 0;
 }
@@ -121,10 +134,12 @@ void showHelp(const char *name)
 {
     printf("Usage: %s [-h/--help]\n"
            "   or: %s [-v/--version]\n"
-           "   or: %s [-x/--index <nvIdx>] [-a/--authHandle <hexHandle>] [-P/--handlePasswd <string>] [-f/--file <filename>]\n"
-           "   or: %s [-x/--index <nvIdx>] [-a/--authHandle <hexHandle>] [-P/--handlePasswd <string>] [-f/--file <filename>]\n"
-           "                    [-X/--passwdInHex]\n"
-           "                    [-p/--port <port>] [-d/--dbg <dbgLevel>]\n"
+           "   or: %s [-x/--index <nvIdx>] [-a/--authHandle <hexHandle>] [-s/--size <size>] [-t/--attribute <attributeDWord>]\n"
+           "                     [-P/--handlePasswd <string>] [-I/--indexPasswd <string>]\n"
+           "   or: %s [-x/--index <nvIdx>] [-a/--authHandle <hexHandle>] [-s/--size <size>] [-t/--attribute <attributeDWord>]\n"
+           "                     [-P/--handlePasswd <string>] [-I/--indexPasswd <string>]\n"
+           "                     [-X/--passwdInHex]\n"
+           "                     [-p/--port <port>] [-d/--dbg <dbgLevel>]\n"
            "\nwhere:\n\n"
            "   -h/--help                       display this help and exit.\n"
            "   -v/--version                    display version information and exit.\n"
@@ -132,9 +147,10 @@ void showHelp(const char *name)
            "   -a/--authHandle <hexHandle>     specifies the handle used to authorize:\n"
            "                                     0x40000001 (TPM_RH_OWNER)\n"
            "                                     0x4000000C (TPM_RH_PLATFORM)\n"
-           "                                     {NV_INDEX_FIRST:NV_INDEX_LAST}\n"
+           "   -s/--size <size>                specifies the size of data area.\n"
+           "   -t/--attribute <attributeDWord> specifies the value of attribute in publicInfo struct (need calculate outside).\n"
            "   -P/--handlePasswd <string>      specifies the password of authHandle.\n"
-           "   -f/--file <filename>            specifies the nv data file.\n"
+           "   -I/--indexPasswd <string>       specifies the password of NV Index when created.\n"
            "   -X/--passwdInHex                passwords given by any options are hex format.\n"
            "   -p/--port <port>                specifies the port number, default:%d, optional\n"
            "   -d/--dbg <dbgLevel>             specifies level of debug messages, optional:\n"
@@ -143,15 +159,14 @@ void showHelp(const char *name)
            "                                     2 (resource manager send/receive byte streams)\n"
            "                                     3 (resource manager tables)\n"
            "\nexample:\n"
-           "   %s -x 0x1500016 -a 0x40000001 -P passwd -f nv.data\n"
-           "   %s -x 0x1500016 -a 0x40000001 -P 1a1b1c -X -f nv.data\n"
+           "   %s -x 0x1500016 -a 0x40000001 -s 32 -t 0x2000A\n"
+           "   %s -x 0x1500016 -a 0x40000001 -s 32 -t 0x2000A -I 1a1b1c -X\n"
            , name, name, name, name, DEFAULT_RESMGR_TPM_PORT, name, name);
 }
 
 int main(int argc, char* argv[])
 {
     int opt;
-    char type[100] = "local";
     char hostName[200] = DEFAULT_HOSTNAME;
     int port = DEFAULT_RESMGR_TPM_PORT;
     int returnVal = 0;
@@ -159,8 +174,10 @@ int main(int argc, char* argv[])
     struct option sOpts[] = {
         { "index"       , required_argument, NULL, 'x' },
         { "authHandle"  , required_argument, NULL, 'a' },
-        { "file"        , required_argument, NULL, 'f' },
+        { "size"        , required_argument, NULL, 's' },
+        { "attribute"   , required_argument, NULL, 't' },
         { "handlePasswd", required_argument, NULL, 'P' },
+        { "indexPasswd" , required_argument, NULL, 'I' },
         { "passwdInHex" , no_argument,       NULL, 'X' },
         { "port"        , required_argument, NULL, 'p' },
         { "dbg"         , required_argument, NULL, 'd' },
@@ -169,7 +186,7 @@ int main(int argc, char* argv[])
         { NULL          , no_argument,       NULL,  0  },
     };
 
-    if(argc == 1)
+    if( argc == 1)
     {
         showHelp(argv[0]);
         return 0;
@@ -181,101 +198,95 @@ int main(int argc, char* argv[])
         return -1;
     }
 
-    while ( ( opt = getopt_long( argc, argv, "x:a:f:P:Xp:d:hv", sOpts, NULL ) ) != -1 )
+    while ( ( opt = getopt_long( argc, argv, "x:a:s:t:P:I:Xp:d:hv", sOpts, NULL ) ) != -1 )
     {
         switch ( opt ) {
         case 'h':
         case '?':
             showHelp(argv[0]);
-            return 0;
+            return -2;
         case 'v':
             showVersion(argv[0]);
-            return 0;
+            return -3;
 
         case 'x':
             if( getSizeUint32Hex(optarg, &nvIndex) != 0 )
-            {
-                return -2;
-            }
+                return -4;
             break;
 
         case 'a':
             if( getSizeUint32Hex(optarg, &authHandle) != 0 )
-            {
-                return -3;
-            }
+                return -5;
+            break;
+
+        case 's':
+            if( getSizeUint32(optarg, &size) != 0 )
+                return -6;
+            break;
+
+        case 't':
+            if( getSizeUint32Hex(optarg, &nvAttribute) != 0 )
+                return -7;
             break;
 
         case 'P':
             if( optarg == NULL || (strlen(optarg) >= sizeof(TPMU_HA)) )
             {
                 printf("\nPlease input the handle password(optional,no more than %d characters).\n", (int)sizeof(TPMU_HA)-1);
-                return -4;
+                return -8;
             }
-            safeStrNCpy(&handlePasswd[0], optarg, sizeof(handlePasswd));
+            safeStrNCpy(handlePasswd, optarg, sizeof(handlePasswd));
             break;
 
-        case 'f':
-            if( optarg == NULL )
+        case 'I':
+            if( optarg == NULL || (strlen(optarg) >= sizeof(TPMU_HA)) )
             {
-                printf("\nPlease input the nv data file.\n");
-                return -5;
+                printf("\nPlease input the index password(optional,no more than %d characters).\n", (int)sizeof(TPMU_HA)-1);
+                return -9;
             }
-            safeStrNCpy(&fileName[0], optarg, sizeof(fileName));
+            safeStrNCpy(indexPasswd, optarg, sizeof(indexPasswd));
             break;
-
         case 'X':
             hexPasswd = true;
             break;
-
-         case 'p':
+        case 'p':
             if( getPort(optarg, &port) )
             {
                 printf("Incorrect port number.\n");
-                return -6;
+                return -10;
             }
             break;
         case 'd':
             if( getDebugLevel(optarg, &debugLevel) )
             {
                 printf("Incorrect debug level.\n");
-                return -7;
+                return -11;
             }
             break;
-
-        default:
-            showArgMismatch(argv[0]);
-            return -8;
         }
     }
 
     if( nvIndex == 0 )
     {
-        printf("You must provide an index (!= 0) for the NVRAM area.\n");
-        return -9;
+        printf("\nYou must provide an index (!= 0) for the NV area.\n");
+        return -12;
     }
 
     if( authHandle == 0 )
     {
-        printf("You must provide an right auth handle for this operation.\n");
-        return -10;
-    }
-
-    dataSize = MAX_NV_INDEX_SIZE;
-    if(loadDataFromFile(fileName, nvBuffer, &dataSize))
-    {
-        printf("Failed to read data from %s\n", fileName);
-        return -11;
+        printf("\nYou must provide an right auth handle for this operation.\n");
+        showHelp(argv[0]);
+        return -13;
     }
 
     prepareTest(hostName, port, debugLevel);
 
-    returnVal = nvWrite();
+    returnVal = nvSpaceDefine();
 
     finishTest();
 
     if(returnVal)
-        return -12;
+        return -14;
 
     return 0;
 }
