+++ src/tpm2_nvreadlock.c
@@ -31,6 +31,7 @@
 //**********************************************************************;
 
 #include <stdarg.h>
+#include <stdbool.h>
 
 #include <stdlib.h>
 #include <stdio.h>
