+++ src/tpm2_nvrelease.c
@@ -30,6 +30,7 @@
 //**********************************************************************;
 
 #include <stdarg.h>
+#include <stdbool.h>
 
 #include <stdlib.h>
 #include <stdio.h>
@@ -43,33 +44,21 @@
 #include "common.h"
 
 int debugLevel = 0;
-UINT32 nvIndex = 0;
-UINT32 authHandle = TPM_RH_PLATFORM;
-UINT32 size = 0;
-UINT32 nvAttribute = 0;
+UINT32 nvIndex;
+UINT32 authHandle;
 char handlePasswd[sizeof(TPMU_HA)];
-char indexPasswd[sizeof(TPMU_HA)];
 bool hexPasswd = false;
 
-int nvSpaceDefine()
+int nvSpaceRelease()
 {
     UINT32 rval;
-    TPM2B_AUTH nvAuth;
-    TPM2B_NV_PUBLIC publicInfo;
     TPMS_AUTH_COMMAND sessionData;
-    TPMS_AUTH_RESPONSE sessionDataOut;
     TSS2_SYS_CMD_AUTHS sessionsData;
-    TSS2_SYS_RSP_AUTHS sessionsDataOut;
+    TPMS_AUTH_COMMAND *sessionDataArray[1];
 
-    TPMS_AUTH_COMMAND *sessionDataArray[1] = { &sessionData };
-    TPMS_AUTH_RESPONSE *sessionDataOutArray[1] = { &sessionDataOut };
-
-    sessionsDataOut.rspAuths = &sessionDataOutArray[0];
+    sessionDataArray[0] = &sessionData;
     sessionsData.cmdAuths = &sessionDataArray[0];
-
-    sessionsDataOut.rspAuthsCount = 1;
     sessionsData.cmdAuthsCount = 1;
-
     sessionData.sessionHandle = TPM_RS_PW;
     sessionData.nonce.t.size = 0;
     sessionData.hmac.t.size = 0;
@@ -91,40 +80,13 @@ int nvSpaceDefine()
         }
     }
 
-    publicInfo.t.size = sizeof( TPMI_RH_NV_INDEX ) +
-            sizeof( TPMI_ALG_HASH ) + sizeof( TPMA_NV ) + sizeof( UINT16) +
-            sizeof( UINT16 );
-    publicInfo.t.nvPublic.nvIndex = nvIndex;
-    publicInfo.t.nvPublic.nameAlg = TPM_ALG_SHA256;
-
-    // Now set the attributes.
-    publicInfo.t.nvPublic.attributes.val = nvAttribute;
-    publicInfo.t.nvPublic.authPolicy.t.size = 0;
-    publicInfo.t.nvPublic.dataSize = size;
-
-    nvAuth.t.size = 0;
-    if (strlen(indexPasswd) > 0 && !hexPasswd)
-    {
-        nvAuth.t.size = strlen(indexPasswd);
-        memcpy( &nvAuth.t.buffer[0], indexPasswd, nvAuth.t.size );
-    }
-    else if (strlen(indexPasswd) > 0 && hexPasswd)
-    {
-        nvAuth.t.size = sizeof(nvAuth) - 2;
-        if (hex2ByteStructure(indexPasswd,&nvAuth.t.size,nvAuth.t.buffer) != 0)
-        {
-            printf( "Failed to convert Hex format password for indexPasswd.\n");
-            return -1;
-        }
-    }
-
-    rval = Tss2_Sys_NV_DefineSpace( sysContext, authHandle, &sessionsData, &nvAuth, &publicInfo, &sessionsDataOut );
+    rval = Tss2_Sys_NV_UndefineSpace( sysContext, authHandle, nvIndex, &sessionsData, 0 );
     if(rval != TPM_RC_SUCCESS)
     {
-        printf( "\nFailed to define NV area at index 0x%x (%d).Error:0x%x\n", nvIndex, nvIndex, rval );
+        printf( "Failed to release NV area at index 0x%x (%d).Error:0x%x\n", nvIndex, nvIndex, rval );
         return -1;
     }
-    printf( "Success to define NV area at index 0x%x (%d).\n", nvIndex, nvIndex);
+    printf( "Success to release NV area at index 0x%x (%d).\n", nvIndex, nvIndex);
 
     return 0;
 }
@@ -133,12 +95,10 @@ void showHelp(const char *name)
 {
     printf("Usage: %s [-h/--help]\n"
            "   or: %s [-v/--version]\n"
-           "   or: %s [-x/--index <nvIdx>] [-a/--authHandle <hexHandle>] [-s/--size <size>] [-t/--attribute <attributeDWord>]\n"
-           "                     [-P/--handlePasswd <string>] [-I/--indexPasswd <string>]\n"
-           "   or: %s [-x/--index <nvIdx>] [-a/--authHandle <hexHandle>] [-s/--size <size>] [-t/--attribute <attributeDWord>]\n"
-           "                     [-P/--handlePasswd <string>] [-I/--indexPasswd <string>]\n"
-           "                     [-X/--passwdInHex]\n"
-           "                     [-p/--port <port>] [-d/--dbg <dbgLevel>]\n"
+           "   or: %s [-x/--index <nvIdx>] [-a/--authHandle <hexHandle>] [-P/--handlePasswd <string>]\n"
+           "   or: %s [-x/--index <nvIdx>] [-a/--authHandle <hexHandle>] [-P/--handlePasswd <string>]\n"
+           "                      [-X/--passwdInHex]\n"
+           "                      [-p/--port <port>] [-d/--dbg <dbgLevel>]\n"
            "\nwhere:\n\n"
            "   -h/--help                       display this help and exit.\n"
            "   -v/--version                    display version information and exit.\n"
@@ -146,10 +106,7 @@ void showHelp(const char *name)
            "   -a/--authHandle <hexHandle>     specifies the handle used to authorize:\n"
            "                                     0x40000001 (TPM_RH_OWNER)\n"
            "                                     0x4000000C (TPM_RH_PLATFORM)\n"
-           "   -s/--size <size>                specifies the size of data area.\n"
-           "   -t/--attribute <attributeDWord> specifies the value of attribute in publicInfo struct (need calculate outside).\n"
            "   -P/--handlePasswd <string>      specifies the password of authHandle.\n"
-           "   -I/--indexPasswd <string>       specifies the password of NV Index when created.\n"
            "   -X/--passwdInHex                passwords given by any options are hex format.\n"
            "   -p/--port <port>                specifies the port number, default:%d, optional\n"
            "   -d/--dbg <dbgLevel>             specifies level of debug messages, optional:\n"
@@ -158,8 +115,8 @@ void showHelp(const char *name)
            "                                     2 (resource manager send/receive byte streams)\n"
            "                                     3 (resource manager tables)\n"
            "\nexample:\n"
-           "   %s -x 0x1500016 -a 0x40000001 -s 32 -t 0x2000A\n"
-           "   %s -x 0x1500016 -a 0x40000001 -s 32 -t 0x2000A -I 1a1b1c -X\n"
+           "   %s -x 0x1500016 -a 0x40000001 -P passwd\n"
+           "   %s -x 0x1500016 -a 0x40000001 -P 1a1b1c -X"
            , name, name, name, name, DEFAULT_RESMGR_TPM_PORT, name, name);
 }
 
@@ -170,13 +127,11 @@ int main(int argc, char* argv[])
     int port = DEFAULT_RESMGR_TPM_PORT;
     int returnVal = 0;
 
+
     struct option sOpts[] = {
         { "index"       , required_argument, NULL, 'x' },
         { "authHandle"  , required_argument, NULL, 'a' },
-        { "size"        , required_argument, NULL, 's' },
-        { "attribute"   , required_argument, NULL, 't' },
         { "handlePasswd", required_argument, NULL, 'P' },
-        { "indexPasswd" , required_argument, NULL, 'I' },
         { "passwdInHex" , no_argument,       NULL, 'X' },
         { "port"        , required_argument, NULL, 'p' },
         { "dbg"         , required_argument, NULL, 'd' },
@@ -185,7 +140,7 @@ int main(int argc, char* argv[])
         { NULL          , no_argument,       NULL,  0  },
     };
 
-    if( argc == 1)
+    if( argc == 1 )
     {
         showHelp(argv[0]);
         return 0;
@@ -197,53 +152,38 @@ int main(int argc, char* argv[])
         return -1;
     }
 
-    while ( ( opt = getopt_long( argc, argv, "x:a:s:t:P:I:Xp:d:hv", sOpts, NULL ) ) != -1 )
+    while ( ( opt = getopt_long( argc, argv, "x:a:P:Xp:d:hv", sOpts, NULL ) ) != -1 )
     {
         switch ( opt ) {
         case 'h':
         case '?':
             showHelp(argv[0]);
-            return -2;
+            return 0;
         case 'v':
             showVersion(argv[0]);
-            return -3;
+            return 0;
 
         case 'x':
             if( getSizeUint32Hex(optarg, &nvIndex) != 0 )
-                return -4;
+            {
+                return -2;
+            }
             break;
 
         case 'a':
             if( getSizeUint32Hex(optarg, &authHandle) != 0 )
-                return -5;
-            break;
-
-        case 's':
-            if( getSizeUint32(optarg, &size) != 0 )
-                return -6;
-            break;
-
-        case 't':
-            if( getSizeUint32Hex(optarg, &nvAttribute) != 0 )
-                return -7;
+            {
+                return -3;
+            }
             break;
 
         case 'P':
             if( optarg == NULL || (strlen(optarg) >= sizeof(TPMU_HA)) )
             {
                 printf("\nPlease input the handle password(optional,no more than %d characters).\n", (int)sizeof(TPMU_HA)-1);
-                return -8;
+                return -4;
             }
             safeStrNCpy(handlePasswd, optarg, sizeof(handlePasswd));
-            break;
-
-        case 'I':
-            if( optarg == NULL || (strlen(optarg) >= sizeof(TPMU_HA)) )
-            {
-                printf("\nPlease input the index password(optional,no more than %d characters).\n", (int)sizeof(TPMU_HA)-1);
-                return -9;
-            }
-            safeStrNCpy(indexPasswd, optarg, sizeof(indexPasswd));
             break;
         case 'X':
             hexPasswd = true;
@@ -252,40 +192,43 @@ int main(int argc, char* argv[])
             if( getPort(optarg, &port) )
             {
                 printf("Incorrect port number.\n");
-                return -10;
+                return -5;
             }
             break;
         case 'd':
             if( getDebugLevel(optarg, &debugLevel) )
             {
                 printf("Incorrect debug level.\n");
-                return -11;
+                return -6;
             }
             break;
+
+        default:
+            showArgMismatch(argv[0]);
+            return -7;
         }
     }
 
     if( nvIndex == 0 )
     {
-        printf("\nYou must provide an index (!= 0) for the NV area.\n");
-        return -12;
+        printf("\nYou must provide an index (!= 0) for the NVRAM area.\n");
+        return -8;
     }
 
     if( authHandle == 0 )
     {
         printf("\nYou must provide an right auth handle for this operation.\n");
-        showHelp(argv[0]);
-        return -13;
+        return -9;
     }
 
     prepareTest(hostName, port, debugLevel);
 
-    returnVal = nvSpaceDefine();
+    returnVal = nvSpaceRelease();
 
     finishTest();
 
     if(returnVal)
-        return -14;
+        return -10;
 
     return 0;
 }
