+++ src/tpm2_nvwrite.c
@@ -37,27 +37,40 @@
 #include <limits.h>
 #include <ctype.h>
 #include <getopt.h>
+#include <stdbool.h>
 
 #include <sapi/tpm20.h>
 #include <tcti/tcti_socket.h>
 #include "common.h"
 
 int debugLevel = 0;
-UINT32 nvIndex;
-UINT32 authHandle;
+UINT32 nvIndex = 0;
+UINT32 authHandle = TPM_RH_PLATFORM;
+UINT16 dataSize = 0;
+unsigned char nvBuffer[MAX_NV_INDEX_SIZE];
+char fileName[PATH_MAX];
 char handlePasswd[sizeof(TPMU_HA)];
 bool hexPasswd = false;
 
-int nvSpaceRelease()
+int nvWrite()
 {
     UINT32 rval;
     TPMS_AUTH_COMMAND sessionData;
+    TPMS_AUTH_RESPONSE sessionDataOut;
     TSS2_SYS_CMD_AUTHS sessionsData;
-    TPMS_AUTH_COMMAND *sessionDataArray[1];
+    TSS2_SYS_RSP_AUTHS sessionsDataOut;
+    UINT16 i, offset = 0;
+    TPM2B_MAX_NV_BUFFER nvWriteData;
 
-    sessionDataArray[0] = &sessionData;
+    TPMS_AUTH_COMMAND *sessionDataArray[1] = { &sessionData };
+    TPMS_AUTH_RESPONSE *sessionDataOutArray[1] = { &sessionDataOut };
+
+    sessionsDataOut.rspAuths = &sessionDataOutArray[0];
     sessionsData.cmdAuths = &sessionDataArray[0];
+
+    sessionsDataOut.rspAuthsCount = 1;
     sessionsData.cmdAuthsCount = 1;
+
     sessionData.sessionHandle = TPM_RS_PW;
     sessionData.nonce.t.size = 0;
     sessionData.hmac.t.size = 0;
@@ -79,13 +92,28 @@ int nvSpaceRelease()
         }
     }
 
-    rval = Tss2_Sys_NV_UndefineSpace( sysContext, authHandle, nvIndex, &sessionsData, 0 );
-    if(rval != TPM_RC_SUCCESS)
+    while (dataSize > 0)
     {
-        printf( "Failed to release NV area at index 0x%x (%d).Error:0x%x\n", nvIndex, nvIndex, rval );
-        return -1;
+        nvWriteData.t.size = dataSize > MAX_NV_BUFFER_SIZE ? MAX_NV_BUFFER_SIZE : dataSize;
+        printf("\nThe data(size=%d) to be written:\n", nvWriteData.t.size);
+        for( i = 0; i<(int)nvWriteData.t.size; i++ )
+        {
+            nvWriteData.t.buffer[i] = nvBuffer[offset+i];
+            printf("%02x ", nvBuffer[offset+i]);
+        }
+        printf("\n\n");
+
+        rval = Tss2_Sys_NV_Write( sysContext, authHandle, nvIndex, &sessionsData, &nvWriteData, offset, &sessionsDataOut );
+        if(rval != TSS2_RC_SUCCESS)
+        {
+            printf( "Failed to write NV area at index 0x%x (%d) offset 0x%x. Error:0x%x\n", nvIndex, nvIndex, offset, rval );
+            return -1;
+        }
+        printf( "Success to write NV area at index 0x%x (%d) offset 0x%x.\n", nvIndex, nvIndex, offset );
+
+        dataSize -= nvWriteData.t.size;
+        offset += nvWriteData.t.size;        
     }
-    printf( "Success to release NV area at index 0x%x (%d).\n", nvIndex, nvIndex);
 
     return 0;
 }
@@ -94,10 +122,10 @@ void showHelp(const char *name)
 {
     printf("Usage: %s [-h/--help]\n"
            "   or: %s [-v/--version]\n"
-           "   or: %s [-x/--index <nvIdx>] [-a/--authHandle <hexHandle>] [-P/--handlePasswd <string>]\n"
-           "   or: %s [-x/--index <nvIdx>] [-a/--authHandle <hexHandle>] [-P/--handlePasswd <string>]\n"
-           "                      [-X/--passwdInHex]\n"
-           "                      [-p/--port <port>] [-d/--dbg <dbgLevel>]\n"
+           "   or: %s [-x/--index <nvIdx>] [-a/--authHandle <hexHandle>] [-P/--handlePasswd <string>] [-f/--file <filename>]\n"
+           "   or: %s [-x/--index <nvIdx>] [-a/--authHandle <hexHandle>] [-P/--handlePasswd <string>] [-f/--file <filename>]\n"
+           "                    [-X/--passwdInHex]\n"
+           "                    [-p/--port <port>] [-d/--dbg <dbgLevel>]\n"
            "\nwhere:\n\n"
            "   -h/--help                       display this help and exit.\n"
            "   -v/--version                    display version information and exit.\n"
@@ -105,7 +133,9 @@ void showHelp(const char *name)
            "   -a/--authHandle <hexHandle>     specifies the handle used to authorize:\n"
            "                                     0x40000001 (TPM_RH_OWNER)\n"
            "                                     0x4000000C (TPM_RH_PLATFORM)\n"
+           "                                     {NV_INDEX_FIRST:NV_INDEX_LAST}\n"
            "   -P/--handlePasswd <string>      specifies the password of authHandle.\n"
+           "   -f/--file <filename>            specifies the nv data file.\n"
            "   -X/--passwdInHex                passwords given by any options are hex format.\n"
            "   -p/--port <port>                specifies the port number, default:%d, optional\n"
            "   -d/--dbg <dbgLevel>             specifies level of debug messages, optional:\n"
@@ -114,22 +144,23 @@ void showHelp(const char *name)
            "                                     2 (resource manager send/receive byte streams)\n"
            "                                     3 (resource manager tables)\n"
            "\nexample:\n"
-           "   %s -x 0x1500016 -a 0x40000001 -P passwd\n"
-           "   %s -x 0x1500016 -a 0x40000001 -P 1a1b1c -X"
+           "   %s -x 0x1500016 -a 0x40000001 -P passwd -f nv.data\n"
+           "   %s -x 0x1500016 -a 0x40000001 -P 1a1b1c -X -f nv.data\n"
            , name, name, name, name, DEFAULT_RESMGR_TPM_PORT, name, name);
 }
 
 int main(int argc, char* argv[])
 {
     int opt;
+    char type[100] = "local";
     char hostName[200] = DEFAULT_HOSTNAME;
     int port = DEFAULT_RESMGR_TPM_PORT;
     int returnVal = 0;
 
-
     struct option sOpts[] = {
         { "index"       , required_argument, NULL, 'x' },
         { "authHandle"  , required_argument, NULL, 'a' },
+        { "file"        , required_argument, NULL, 'f' },
         { "handlePasswd", required_argument, NULL, 'P' },
         { "passwdInHex" , no_argument,       NULL, 'X' },
         { "port"        , required_argument, NULL, 'p' },
@@ -139,7 +170,7 @@ int main(int argc, char* argv[])
         { NULL          , no_argument,       NULL,  0  },
     };
 
-    if( argc == 1 )
+    if(argc == 1)
     {
         showHelp(argv[0]);
         return 0;
@@ -151,7 +182,7 @@ int main(int argc, char* argv[])
         return -1;
     }
 
-    while ( ( opt = getopt_long( argc, argv, "x:a:P:Xp:d:hv", sOpts, NULL ) ) != -1 )
+    while ( ( opt = getopt_long( argc, argv, "x:a:f:P:Xp:d:hv", sOpts, NULL ) ) != -1 )
     {
         switch ( opt ) {
         case 'h':
@@ -182,52 +213,70 @@ int main(int argc, char* argv[])
                 printf("\nPlease input the handle password(optional,no more than %d characters).\n", (int)sizeof(TPMU_HA)-1);
                 return -4;
             }
-            safeStrNCpy(handlePasswd, optarg, sizeof(handlePasswd));
+            safeStrNCpy(&handlePasswd[0], optarg, sizeof(handlePasswd));
             break;
+
+        case 'f':
+            if( optarg == NULL )
+            {
+                printf("\nPlease input the nv data file.\n");
+                return -5;
+            }
+            safeStrNCpy(&fileName[0], optarg, sizeof(fileName));
+            break;
+
         case 'X':
             hexPasswd = true;
             break;
-        case 'p':
+
+         case 'p':
             if( getPort(optarg, &port) )
             {
                 printf("Incorrect port number.\n");
-                return -5;
+                return -6;
             }
             break;
         case 'd':
             if( getDebugLevel(optarg, &debugLevel) )
             {
                 printf("Incorrect debug level.\n");
-                return -6;
+                return -7;
             }
             break;
 
         default:
             showArgMismatch(argv[0]);
-            return -7;
+            return -8;
         }
     }
 
     if( nvIndex == 0 )
     {
-        printf("\nYou must provide an index (!= 0) for the NVRAM area.\n");
-        return -8;
+        printf("You must provide an index (!= 0) for the NVRAM area.\n");
+        return -9;
     }
 
     if( authHandle == 0 )
     {
-        printf("\nYou must provide an right auth handle for this operation.\n");
-        return -9;
+        printf("You must provide an right auth handle for this operation.\n");
+        return -10;
+    }
+
+    dataSize = MAX_NV_INDEX_SIZE;
+    if(loadDataFromFile(fileName, nvBuffer, &dataSize))
+    {
+        printf("Failed to read data from %s\n", fileName);
+        return -11;
     }
 
     prepareTest(hostName, port, debugLevel);
 
-    returnVal = nvSpaceRelease();
+    returnVal = nvWrite();
 
     finishTest();
 
     if(returnVal)
-        return -10;
+        return -12;
 
     return 0;
 }
