+++ src/tpm2_takeownership.c
@@ -30,6 +30,7 @@
 //**********************************************************************;
 
 #include <stdarg.h>
+#include <stdbool.h>
 
 #include <stdlib.h>
 #include <stdio.h>
@@ -432,4 +433,3 @@ int main(int argc, char *argv[])
 
     return 0;
 }
-
